tests/wpt/tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.put.dirty.rect2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.imageData.put.dirty.rect2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.put.dirty.rect2</h1>
<p class="desc">putImageData() only modifies areas inside the dirty rectangle, using x and y</p>


<script>
var t = async_test("putImageData() only modifies areas inside the dirty rectangle, using x and y");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50)
  ctx.fillStyle = '#0f0';
  ctx.fillRect(60, 30, 20, 20)

  var imgdata = ctx.getImageData(0, 0, 100, 50);

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50)
  ctx.fillStyle = '#f00';
  ctx.fillRect(40, 20, 20, 20)
  ctx.putImageData(imgdata, -20, -10, 60, 30, 20, 20);

  _assertPixelApprox(canvas, 50,25, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 35,25, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 65,25, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 50,15, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 50,45, 0,255,0,255, 2);
  t.done();

});
</script>
